fix(sdk): simplify video read handling
