+++ libs/deepagents/tests/unit_tests/test_middleware.py
@@ -1,5 +1,3 @@
-import base64
-import json
 import mimetypes
 import time
 from unittest.mock import MagicMock, patch
@@ -16,7 +14,6 @@ from langchain_core.messages import (
     ToolCall,
     ToolMessage,
 )
-from langchain_core.messages.utils import convert_to_openai_messages
 from langgraph.graph.message import REMOVE_ALL_MESSAGES
 from langgraph.store.memory import InMemoryStore
 from langgraph.types import Command
@@ -1470,299 +1467,6 @@ class TestFilesystemMiddleware:
 
         assert result == tool_message
 
-    def test_read_file_video_routes_to_frame_extractor(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """Video reads route through `_video.extract_video_frames`, not the generic base64 branch.
-
-        The middleware must interpret `offset`/`limit` as seconds for video
-        reads, return interleaved text+image content blocks, and never leak the
-        raw video bytes to the model.
-        """
-        raw_bytes = b"\x00\x01\x02 fake video bytes"
-        sentinel = [
-            {"type": "text", "text": "Frame at t=00:00:00.000"},
-            {"type": "image", "base64": "AAAA", "mime_type": "image/jpeg"},
-            {"type": "text", "text": "Frame at t=00:00:02.000"},
-            {"type": "image", "base64": "BBBB", "mime_type": "image/jpeg"},
-        ]
-
-        def fake_extract(content, *, offset_seconds, duration_seconds, sampling_rate):  # type: ignore[no-untyped-def]
-            # Capture the call arguments so the test asserts the
-            # offset/limit -> seconds reinterpretation.
-            fake_extract.calls.append(
-                {
-                    "len": len(content),
-                    "offset_seconds": offset_seconds,
-                    "duration_seconds": duration_seconds,
-                    "sampling_rate": sampling_rate,
-                }
-            )
-            return list(sentinel)
-
-        fake_extract.calls = []
-
-        monkeypatch.setattr(filesystem_middleware, "extract_video_frames", fake_extract)
-
-        middleware = FilesystemMiddleware(backend=_video_backend(raw_bytes))
-        state = FilesystemState(messages=[], files={})
-        runtime = _build_runtime(state, "video-read-1")
-        read_file_tool = next(t for t in middleware.tools if t.name == "read_file")
-
-        result = read_file_tool.invoke(
-            {
-                "file_path": "/clips/intro.mkv",
-                "offset": 0,
-                "limit": 30,
-                "runtime": runtime,
-            }
-        )
-
-        assert isinstance(result, Command)
-        messages = result.update["messages"]  # type: ignore[index]
-        assert len(messages) == 2
-        tool_message, media_message = messages
-        assert isinstance(tool_message, ToolMessage)
-        assert tool_message.status == "success"
-        assert tool_message.content == "Read video /clips/intro.mkv: sampled 2 frames. The sampled frames are attached in the following message."
-        assert tool_message.additional_kwargs["read_file_path"] == "/clips/intro.mkv"
-        assert tool_message.additional_kwargs["read_file_frame_count"] == 2
-        assert isinstance(media_message, HumanMessage)
-        assert media_message.additional_kwargs["read_file_media_result"] is True
-        assert media_message.additional_kwargs["read_file_path"] == "/clips/intro.mkv"
-        assert media_message.additional_kwargs["read_file_tool_call_id"] == "video-read-1"
-        # Raw video bytes must NOT leak to the model: the media message contains
-        # extracted frames, not the original base64 from the backend.
-        assert media_message.content == [
-            {"type": "text", "text": "Reading first 30s of /clips/intro.mkv at 0.5 fps."},
-            *sentinel,
-        ]
-        assert b"\x00\x01\x02" not in str(media_message.content).encode()
-
-        openai_messages = convert_to_openai_messages(
-            [
-                HumanMessage(content="read video"),
-                AIMessage(content="", tool_calls=[{"id": "video-read-1", "name": "read_file", "args": {"file_path": "/clips/intro.mkv"}}]),
-                *messages,
-            ]
-        )
-        openai_tool_message = next(message for message in openai_messages if message["role"] == "tool")
-        assert isinstance(openai_tool_message["content"], str)
-        assert all(
-            block.get("type") != "video"
-            for message in openai_messages
-            for block in (message["content"] if isinstance(message.get("content"), list) else [])
-        )
-
-        # The middleware must convert offset/limit to seconds and use the
-        # internal sampling rate.
-        assert len(fake_extract.calls) == 1
-        call = fake_extract.calls[0]
-        assert call["offset_seconds"] == 0
-        assert call["duration_seconds"] == 30
-        assert call["sampling_rate"] == 0.5
-        assert call["len"] == len(raw_bytes)
-
-    @pytest.mark.parametrize(
-        ("tool_input", "expected"),
-        [
-            (
-                {"offset": 12, "limit": 90},
-                {"offset_seconds": 12.0, "duration_seconds": 90.0, "sampling_rate": 0.5},
-            ),
-            (
-                {},
-                {"offset_seconds": 0.0, "duration_seconds": 100.0, "sampling_rate": 0.5},
-            ),
-        ],
-        ids=["explicit-window", "default-window"],
-    )
-    def test_read_file_video_window_forwards_seconds(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-        tool_input: dict[str, int],
-        expected: dict[str, float],
-    ) -> None:
-        """`offset`/`limit` become seconds, with the default window preserved."""
-        captured: dict[str, float] = {}
-
-        def fake_extract(content, *, offset_seconds, duration_seconds, sampling_rate):  # type: ignore[no-untyped-def]  # noqa: ARG001
-            captured.update(offset_seconds=offset_seconds, duration_seconds=duration_seconds, sampling_rate=sampling_rate)
-            return [{"type": "text", "text": "ok"}]
-
-        monkeypatch.setattr(filesystem_middleware, "extract_video_frames", fake_extract)
-        middleware = FilesystemMiddleware(backend=_video_backend())
-        state = FilesystemState(messages=[], files={})
-        runtime = _build_runtime(state, "video-read-window")
-        read_file_tool = next(t for t in middleware.tools if t.name == "read_file")
-
-        read_file_tool.invoke({"file_path": "/c.mp4", "runtime": runtime, **tool_input})
-
-        assert captured == expected
-
-    def test_read_file_video_media_result_ordered_after_parallel_tool_results(self) -> None:
-        """Video frame attachments do not split a provider-required tool result batch."""
-        ai_message = AIMessage(
-            content="",
-            tool_calls=[
-                {"id": "call_video", "name": "read_file", "args": {"file_path": "/c.mp4"}},
-                {"id": "call_ls", "name": "ls", "args": {"path": "/"}},
-            ],
-        )
-        video_tool = ToolMessage(content="sampled frames", name="read_file", tool_call_id="call_video")
-        video_media = HumanMessage(
-            content=[
-                {"type": "text", "text": "Reading first 100s of /c.mp4 at 0.5 fps."},
-                {"type": "image", "base64": "AAAA", "mime_type": "image/jpeg"},
-            ],
-            additional_kwargs={"read_file_media_result": True},
-        )
-        ls_tool = ToolMessage(content="[]", name="ls", tool_call_id="call_ls")
-
-        user_message = HumanMessage(content="read and list")
-        reordered = filesystem_middleware._move_media_results_after_tool_results([user_message, ai_message, video_tool, video_media, ls_tool])
-
-        assert reordered == [user_message, ai_message, video_tool, ls_tool, video_media]
-
-    def test_read_file_video_extraction_error_surfaces_as_error_message(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """PyAV failures and missing-dep errors render as a tool error, not an exception."""
-        monkeypatch.setattr(
-            filesystem_middleware,
-            "extract_video_frames",
-            lambda *_args, **_kwargs: (_ for _ in ()).throw(filesystem_middleware.VideoExtractionError("av not installed")),
-        )
-        middleware = FilesystemMiddleware(backend=_video_backend(b"corrupt"))
-        state = FilesystemState(messages=[], files={})
-        runtime = _build_runtime(state, "video-read-err")
-        read_file_tool = next(t for t in middleware.tools if t.name == "read_file")
-
-        result = read_file_tool.invoke({"file_path": "/c.mp4", "runtime": runtime})
-
-        assert isinstance(result, ToolMessage)
-        assert result.status == "error"
-        assert "av not installed" in result.content  # type: ignore[operator]
-
-    def test_read_file_video_non_positive_limit_rejected(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """A non-positive `limit` is rejected as a tool error, not silently clamped."""
-        called = {"count": 0}
-
-        def fake_extract(*args: "object", **kwargs: "object") -> "list[dict[str, object]]":  # noqa: ARG001
-            called["count"] += 1
-            return [{"type": "text", "text": "ok"}]
-
-        monkeypatch.setattr(filesystem_middleware, "extract_video_frames", fake_extract)
-        middleware = FilesystemMiddleware(backend=_video_backend())
-        state = FilesystemState(messages=[], files={})
-        runtime = _build_runtime(state, "video-read-bad-limit")
-        read_file_tool = next(t for t in middleware.tools if t.name == "read_file")
-
-        result = read_file_tool.invoke({"file_path": "/c.mp4", "limit": 0, "runtime": runtime})
-
-        assert isinstance(result, ToolMessage)
-        assert result.status == "error"
-        assert "limit must be > 0" in result.content  # type: ignore[operator]
-        assert called["count"] == 0
-
-    def test_read_file_video_payload_size_cap_rejected(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """State/store/custom video payloads are capped after base64 decoding."""
-        called = {"count": 0}
-
-        def fake_extract(*args: "object", **kwargs: "object") -> "list[dict[str, object]]":  # noqa: ARG001
-            called["count"] += 1
-            return [{"type": "text", "text": "ok"}]
-
-        monkeypatch.setattr(filesystem_middleware, "MAX_VIDEO_INPUT_BYTES", 3)
-        monkeypatch.setattr(filesystem_middleware, "extract_video_frames", fake_extract)
-        middleware = FilesystemMiddleware(backend=_video_backend(b"abcd"))
-        state = FilesystemState(messages=[], files={})
-        runtime = _build_runtime(state, "video-read-big-payload")
-        read_file_tool = next(t for t in middleware.tools if t.name == "read_file")
-
-        result = read_file_tool.invoke({"file_path": "/c.mp4", "runtime": runtime})
-
-        assert isinstance(result, ToolMessage)
-        assert result.status == "error"
-        assert "video payload exceeds maximum input size of 3 bytes" in result.content  # type: ignore[operator]
-        assert called["count"] == 0
-
-    def test_user_supplied_video_block_transforms_for_non_video_model(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """Native video blocks are converted to sampled frames for non-video models."""
-        raw_bytes = b"user supplied video"
-        sentinel = [
-            {"type": "text", "text": "Frame at t=00:00:00.000"},
-            {"type": "image", "base64": "AAAA", "mime_type": "image/jpeg"},
-        ]
-        calls: list[dict[str, object]] = []
-
-        def fake_extract(content, *, offset_seconds, duration_seconds, sampling_rate):  # type: ignore[no-untyped-def]
-            calls.append(
-                {
-                    "content": content,
-                    "offset_seconds": offset_seconds,
-                    "duration_seconds": duration_seconds,
-                    "sampling_rate": sampling_rate,
-                }
-            )
-            return list(sentinel)
-
-        monkeypatch.setattr(filesystem_middleware, "extract_video_frames", fake_extract)
-        video_block = _video_content_block(raw_bytes, filename="upload.mp4")
-        messages = [HumanMessage(content=[{"type": "text", "text": "watch this"}, video_block])]
-
-        transformed, mutated = filesystem_middleware._transform_video_blocks_for_model(messages, native_video=False)
-
-        assert mutated is True
-        assert transformed[0] is not messages[0]
-        assert messages[0].content == [{"type": "text", "text": "watch this"}, video_block]
-        assert transformed[0].content == [
-            {"type": "text", "text": "watch this"},
-            {"type": "text", "text": "Reading first 100s of upload.mp4 at 0.5 fps."},
-            *sentinel,
-        ]
-        assert calls == [
-            {
-                "content": raw_bytes,
-                "offset_seconds": 0.0,
-                "duration_seconds": 100.0,
-                "sampling_rate": 0.5,
-            }
-        ]
-        assert video_block["base64"] not in str(transformed[0].content)
-
-    def test_user_supplied_video_block_passes_through_for_video_model(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """Video-capable models keep native video blocks unchanged."""
-        called = {"count": 0}
-
-        def fake_extract(*args: object, **kwargs: object) -> list[dict[str, object]]:  # noqa: ARG001
-            called["count"] += 1
-            return [{"type": "text", "text": "unexpected"}]
-
-        monkeypatch.setattr(filesystem_middleware, "extract_video_frames", fake_extract)
-        messages = [HumanMessage(content=[_video_content_block(b"native", filename="gemini.mov")])]
-        native_video = filesystem_middleware._model_accepts_native_video(_VideoCapableModel("google_genai", "gemini-2.0-flash"))
-
-        transformed, mutated = filesystem_middleware._transform_video_blocks_for_model(messages, native_video=native_video)
-
-        assert native_video is True
-        assert transformed is messages
-        assert mutated is False
-        assert called["count"] == 0
-
-    def test_stringified_video_blocks_are_recovered_for_non_video_model(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """Stringified video tool content is still converted before non-video model calls."""
-        monkeypatch.setattr(
-            filesystem_middleware,
-            "extract_video_frames",
-            lambda *_args, **_kwargs: [{"type": "image", "base64": "AAAA", "mime_type": "image/jpeg"}],
-        )
-        payload = json.dumps([_video_content_block(b"tool video", filename="tool.mp4")])
-        messages = [ToolMessage(content=payload, name="custom_tool", tool_call_id="call_1")]
-
-        transformed, mutated = filesystem_middleware._transform_video_blocks_for_model(messages, native_video=False)
-
-        assert mutated is True
-        assert isinstance(transformed[0].content, list)
-        assert transformed[0].content[0] == {"type": "text", "text": "Reading first 100s of tool.mp4 at 0.5 fps."}
-        assert all(block.get("type") != "video" for block in transformed[0].content)
-
     def test_read_file_image_returns_standard_image_content_block(self):
         """Test image reads return standard image blocks with base64 + mime_type."""
 
@@ -2876,47 +2580,3 @@ class TestBuiltinTruncationTools:
 
         with pytest.raises(ValueError, match="max_execute_timeout must be positive"):
             FilesystemMiddleware(max_execute_timeout=-1)
-
-
-def _build_runtime(state: FilesystemState, tool_call_id: str) -> ToolRuntime:
-    """Build a `ToolRuntime` for the binary/video read_file tests."""
-    return ToolRuntime(
-        state=state,
-        context=None,
-        tool_call_id=tool_call_id,
-        store=None,
-        stream_writer=lambda _: None,
-        config={},
-    )
-
-
-def _video_backend(raw: bytes = b"raw") -> StateBackend:
-    """Backend stub that returns `raw` base64-encoded for any video read."""
-    payload = base64.b64encode(raw).decode("ascii")
-
-    class VideoBackend(StateBackend):
-        def read(self, path, *, offset=0, limit=100):  # type: ignore[override]
-            return ReadResult(file_data={"content": payload, "encoding": "base64"})
-
-    return VideoBackend()
-
-
-def _video_content_block(raw: bytes, *, filename: str) -> dict[str, object]:
-    """Build an inline video content block for request-time transformation tests."""
-    return {
-        "type": "video",
-        "base64": base64.b64encode(raw).decode("ascii"),
-        "mime_type": "video/mp4",
-        "source_metadata": {"filename": filename},
-    }
-
-
-class _VideoCapableModel:
-    """Small model stub exposing the provider metadata used for video capability."""
-
-    def __init__(self, provider: str, model_name: str) -> None:
-        self.model_name = model_name
-        self._provider = provider
-
-    def _get_ls_params(self) -> dict[str, str]:
-        return {"ls_provider": self._provider}
